Prevent integer overflow in Interval countBases method.

+++ src/main/java/htsjdk/samtools/util/Interval.java
@@ -171,7 +171,7 @@ public class Interval implements Comparable<Interval>, Cloneable, Feature {
      */
     public static long countBases(final Collection<Interval> intervals) {
         return intervals.stream()
-                .mapToInt(Interval::length)
+                .mapToLong(Interval::length)
                 .sum();
     }
